Further auth configuration
- JWT & Redis now working, started protecting routes.

+++ packages/api/resolvers/index.js
@@ -31,37 +31,43 @@ const resolvers = {
   Query: {
 
     // Gets Proposals
-    Proposals: async (root, args, context) => {
+    Proposals: async (root, args, { user }) => {
+      if (user === null) throw new Error('Not Authorised!');
       const dbresponse = await findAllProposals();
       return dbresponse;
     },
 
     // Gets Proposals
-    Proposal: async (root, { id }, context) => {
+    Proposal: async (root, { id }, { user }) => {
+      if (user === null) throw new Error('Not Authorised!');
       const dbresponse = await findProposal(id);
       return dbresponse;
     },
 
     // Gets Sections based on Proposal ID
-    Sections: async (root, { id }, context) => {
+    Sections: async (root, { id }, { user }) => {
+      if (user === null) throw new Error('Not Authorised!');
       const dbresponse = await findAllSections(id);
       return dbresponse;
     },
 
     // Gets a Section based on Section ID
-    Section: async (root, { id }, context) => {
+    Section: async (root, { id }, { user }) => {
+      if (user === null) throw new Error('Not Authorised!');
       const dbresponse = await findSection(id);
       return dbresponse;
     },
 
     // Gets Products
-    Products: async (root, { status }, context) => {
+    Products: async (root, { status }, { user }) => {
+      if (user === null) throw new Error('Not Authorised!');
       const dbresponse = await findAllProducts(status);
       return dbresponse;
     },
 
     // Gets Product
-    Product: async (root, { id, status }, context) => {
+    Product: async (root, { id, status }, { user }) => {
+      if (user === null) throw new Error('Not Authorised!');
       const dbresponse = await findProduct(id, status);
       return dbresponse;
     },
@@ -83,56 +89,65 @@ const resolvers = {
   Mutation: {
 
     // Create One Proposal
-    addProposal: async (root, proposal, context) => {
+    addProposal: async (root, proposal, { user }) => {
+      if (user === null) throw new Error('Not Authorised!');
       const dbrespones = await createProposal(proposal);
       return dbrespones;
     },
 
     // Update Proposal
-    updateProposal: async (root, proposal, context) => {
+    updateProposal: async (root, proposal, { user }) => {
+      if (user === null) throw new Error('Not Authorised!');
       const dbrespones = await updateProposal(proposal);
       return dbrespones;
     },
 
     // Delete Proposal
-    deleteProposal: async (root, proposal, context) => {
+    deleteProposal: async (root, proposal, { user }) => {
+      if (user === null) throw new Error('Not Authorised!');
       const dbrespones = await deleteProposal(proposal);
       return dbrespones;
     },
 
     // Add Section
-    addSection: async (root, section, context) => {
+    addSection: async (root, section, { user }) => {
+      if (user === null) throw new Error('Not Authorised!');
       const dbresponse = await createSection(section);
       return (dbresponse);
     },
 
     // Update Section
-    updateSection: async (root, section, context) => {
+    updateSection: async (root, section, { user }) => {
+      if (user === null) throw new Error('Not Authorised!');
       const dbresponse = await updateSection(section);
       return (dbresponse);
     },
 
     // Update Section
-    updateSectionOrder: async (root, { data }, context) => {
+    updateSectionOrder: async (root, { data }, { user }) => {
+      if (user === null) throw new Error('Not Authorised!');
       const { parentId, from, to } = data;
       const dbresponse = await updateSectionOrder(parentId, from, to);
       return (dbresponse);
     },
 
     // Delete Proposal
-    deleteSection: async (root, section, context) => {
+    deleteSection: async (root, section, { user }) => {
+      if (user === null) throw new Error('Not Authorised!');
       const dbresponse = await deleteSection(section);
       return dbresponse;
     },
 
     // Add Product or Multiple Products
-    addProducts: async (root, { products }, context) => {
+    addProducts: async (root, { products }, { user }) => {
+      if (user === null) throw new Error('Not Authorised!');
       const dbresponse = await createProducts(products);
       return dbresponse;
     },
 
     // Update Product
-    updateProduct: async (root, { product }, context) => {
+    updateProduct: async (root, { product }, { user }) => {
+      if (user === null) throw new Error('Not Authorised!');
       const dbresponse = await updateProduct(product);
       return (dbresponse);
     },
@@ -144,7 +159,7 @@ const resolvers = {
     },
 
     // Auth User (Login), Sets Cookie
-    authUser: async (root, data, { res }) => {
+    authUser: async (root, data, { req, res }) => {
       const { token, authorisedUser } = await authUser(data);
       res.cookie('auth', token, { httpOnly: true });
       return (authorisedUser);
